requirements.txt
# Package dependencies
pandas
numpy
pyyaml
jupyter
scikit-learn
ipykernel

requests
openai
pyarrow
streamlit
python-dotenv
